pyproject.toml
[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "default"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.poetry]
name = "afanchatserver"
version = "0.1.0"
description = "AfanChat服务端"
authors = ["flz <f2095522823@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

django-environ="^0.11.2"

django-ninja-extra = "^0.20.2"
django-ninja-jwt = {extras = ["crypto"], version = "^5.2.10"}
ninja-schema = "^0.13.6"

setuptools = "^69.0.2"

channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.1.0"
hiredis = "^2.2.3"

xmltodict = "^0.13.0"

mysqlclient = "^2.2.1"

celery = "^5.3.6"
django-celery-beat = "^2.5.0"
flower = "^2.0.1"
django-celery-results = "^2.5.1"
eventlet = "^0.34.2"

lxml = "^5.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
fake-useragent = "^1.4.0"

pyahocorasick = "^2.0.0"
django-extensions = "^3.2.3"
openai = "0.028"
pytest = "^7.4.4"
pymysql = "^1.1.0"
minio = "^7.2.3"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
